move hier test to base
