+++ sktime/split/base/tests/test_base.py
@@ -1,6 +1,7 @@
 # copyright: sktime developers, BSD-3-Clause License (see LICENSE file)
 """Tests for base splitter class and functionality."""
 
+import numpy as np
 import pandas as pd
 import pytest
 
@@ -51,8 +52,48 @@ def test_split_loc():
         y.loc[test]
 
 
+def test_split_series_hier():
+    """Tests that split works with hierarchical data."""
+    hierarchy_levels = (2, 4)
+    n_instances = np.prod(hierarchy_levels)
+    n = 12
+    y = _make_hierarchical(
+        hierarchy_levels=hierarchy_levels, max_timepoints=n, min_timepoints=n
+    )
+    cv = SlidingWindowSplitter()
+
+    for train, test in cv.split(y):
+        assert isinstance(train, np.ndarray)
+        assert train.ndim == 1
+        assert train.dtype == np.int64
+        assert len(train) == 10 * n_instances
+        assert isinstance(test, np.ndarray)
+        assert test.ndim == 1
+        assert pd.api.types.is_integer_dtype(test.dtype)
+        assert len(test) == 1 * n_instances
+
+    for train, test in cv.split_loc(y):
+        assert isinstance(train, pd.MultiIndex)
+        assert len(train) == 10 * n_instances
+        assert train.isin(y.index).all()
+        assert isinstance(test, pd.MultiIndex)
+        assert len(test) == 1 * n_instances
+        assert test.isin(y.index).all()
+
+    def inst_index(y):
+        return set(y.index.droplevel(-1).unique())
+
+    for train, test in cv.split_series(y):
+        assert isinstance(train, pd.DataFrame)
+        assert len(train) == 10 * n_instances
+        assert isinstance(test, pd.DataFrame)
+        assert len(test) == 1 * n_instances
+        assert inst_index(train) == inst_index(y)
+        assert inst_index(test) == inst_index(y)
+
+
 def test_hierachical_singlewindowsplitter():
-    """Test broadcasting of splitters to hierarchical data.
+    """Test broadcasting of window splitters to hierarchical data.
 
     Also certifies for failure cas in bug #4972.
     """
